requirements.txt
flask==2.3.3
nltk==3.8.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
PyPDF2==3.0.1
python-docx==1.1.0
Werkzeug==3.0.1
gunicorn>=20.0.0